+++ packages/lumx-react/src/components/text-field/TextField.test.tsx
@@ -17,6 +17,7 @@ import partition from 'lodash/partition';
 import userEvent from '@testing-library/user-event';
 
 import { isFocusVisible } from '@lumx/react/utils/browser/isFocusVisible';
+import { createRef } from 'react';
 import { TextField, TextFieldProps } from './TextField';
 
 const CLASSNAME = TextField.className as string;
@@ -54,7 +55,8 @@ describe(`<${TextField.displayName}>`, () => {
 
     describe('Render', () => {
         it('should render defaults', () => {
-            const { element, inputNative } = setup({ id: 'fixedId' });
+            const inputRef = createRef<HTMLInputElement>();
+            const { element, inputNative } = setup({ id: 'fixedId', inputRef });
             expect(element).toBeInTheDocument();
 
             expect(element).not.toHaveClass(`${CLASSNAME}--is-valid`);
@@ -67,13 +69,16 @@ describe(`<${TextField.displayName}>`, () => {
 
             expect(element).toHaveClass(`${CLASSNAME}--theme-light`);
             expect(inputNative.tagName).toBe('INPUT');
+            expect(inputRef.current).toBe(inputNative);
         });
 
         it('should render textarea', () => {
-            const { element, inputNative } = setup({ id: 'fixedId', multiline: true });
+            const inputRef = createRef<HTMLTextAreaElement>();
+            const { element, inputNative } = setup({ id: 'fixedId', multiline: true, inputRef });
             expect(element).toBeInTheDocument();
 
             expect(inputNative.tagName).toBe('TEXTAREA');
+            expect(inputRef.current).toBe(inputNative);
         });
     });
 
@@ -176,6 +181,8 @@ describe(`<${TextField.displayName}>`, () => {
             expect(error).toHaveTextContent('error');
             expect(helper).toHaveTextContent('helper');
             expect(inputNative).toHaveAttribute('aria-describedby', expect.stringContaining('aria-description'));
+            expect(inputNative).toHaveAttribute('aria-describedby', expect.stringContaining('text-field-error-'));
+            expect(inputNative).toHaveAttribute('aria-describedby', expect.stringContaining('text-field-helper-'));
         });
     });